refactor(resources): move the service interfaces to a service.go file
Move the service interfaces to a seperate file to declutter.
