+++ apiserver/internal/handlers/resource/resourcemigration.go
@@ -5,7 +5,6 @@ package resource
 
 import (
 	"context"
-	"io"
 	"net/http"
 	"net/url"
 	"strconv"
@@ -22,33 +21,6 @@ import (
 	internalerrors "github.com/juju/juju/internal/errors"
 	"github.com/juju/juju/rpc/params"
 )
-
-// ApplicationServiceGetter is an interface for getting resource and
-// application service.
-type ApplicationServiceGetter interface {
-	// Application returns the model's application service.
-	Application(*http.Request) (ApplicationService, error)
-}
-
-// ApplicationService is an interface for the application domain service.
-type ApplicationService interface {
-	// GetApplicationIDByName returns an application ID by application Name. It
-	// returns an error if the application can not be found by the Name.
-	GetApplicationIDByName(ctx context.Context, name string) (coreapplication.ID, error)
-
-	// GetApplicationIDByUnitName returns the application ID for the named unit.
-	GetApplicationIDByUnitName(ctx context.Context, unitName coreunit.Name) (coreapplication.ID, error)
-
-	// GetUnitUUID returns the UUID for the named unit.
-	GetUnitUUID(ctx context.Context, unitName coreunit.Name) (coreunit.UUID, error)
-}
-
-// Resources represents the methods required to migrate a
-// resource blob.
-type Resources interface {
-	SetUnitResource(string, string, charmresource.Resource) (resource.Resource, error)
-	SetResource(string, string, charmresource.Resource, io.Reader, bool) (resource.Resource, error)
-}
 
 // resourcesMigrationUploadHandler handles resources uploads for model migrations.
 type resourcesMigrationUploadHandler struct {
@@ -176,7 +148,7 @@ func (h *resourcesMigrationUploadHandler) processPost(
 			return empty, internalerrors.Errorf("extracting resource details from request: %w", err)
 		}
 
-		reader, err := h.validator.Validate(r.Body, details.fingerprint.String(), details.size)
+		reader, err := h.validator.ValidateAndStoreReader(r.Body, details.fingerprint.String(), details.size)
 		if err != nil {
 			return empty, internalerrors.Errorf("validating resource size and hash: %w", err)
 		}
